requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.38.0
Pillow>=10.2.0
numpy>=1.26.0
huggingface_hub>=0.21.0
matplotlib>=3.7.0
soundfile>=0.12.0
jiwer>=3.0.0
edge-tts>=6.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
boto3>=1.34.0
google-cloud-vision>=3.12.1
pg8000==1.31.5
redis==7.1.0
